feat(core): add NoAuthentication services extension
Trelnex.Core.Api was built assuming authentication and authorization. So it assumes the authentication and authorization services are injected for the web application. And the IServiceProvider is injected for Swagger. If those services are not injected, the web application will fail.

Introduce the NoAuthentication services extension. This will inject those required services, not no actual authentication or authorization is configured.